+++ app/services/policies/copy_policy.rb
@@ -22,7 +22,7 @@ module Policies
 
     def duplicate(source_resource)
       target_resource = source_resource.dup
-      target_resource.path = @copy_to + sub_path(@copy_from, source_resource.path)
+      target_resource.pretty_path = @copy_to + sub_path(@copy_from, source_resource.pretty_path)
       target_resource.save
 
       target_resource