+++ spec/factories/patients.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 FactoryGirl.define do
   factory :patient do
-    case_number { 'c' + Faker::Number.number(6).to_s }
+    case_number { 'c' + Faker::Number.unique.number(6).to_s }
 
     trait :with_pipeline do
       case_number '9900'