+++ app/services/policies/create_policy.rb
@@ -9,7 +9,7 @@ module Policies
 
     def call
       Resource.transaction do
-        resource = Resource.create(service: @service,
+        resource = Resource.create!(service: @service,
                                    pretty_path: @json_body['path'],
                                    resource_type: :local)
         create_access_policies(resource)