+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 ### Changed
 - File browser template for multiple embed mode created (@dharezlak)
 - WebDAV synchronisation takes care for pipeline-related and patient input files (@Nuanda)
+- Computations are executed in the pipeline scope (@mkasztelnik)
 
 ### Deprecated