+++ CHANGELOG.md
@@ -29,6 +29,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 - Move and copy policy API calls handle well resources with escaped characters (@dharezlak)
 - Copy and move destinations for policy management API treated as pretty paths for proper processing (@dharezlak)
 - JS resources required for the OFF viewer are loaded in the file browser direct embed mode (@dharezlak)
+- Resource pretty path decodes and encodes values according to URI restrictions (@dharezlak)
 
 ### Security