+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 - Computations are executed in the pipeline scope (@mkasztelnik)
 - Move FileStore configuration into one place (@mkasztelnik)
 - Make patient `case_number` factory field unique (@mkasztelnik)
+- Redirect user back to previous page after proxy is regenerated successfully (@mkasztelnik)
 
 ### Deprecated