+++ CHANGELOG.md
@@ -32,6 +32,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 - Copy and move destinations for policy management API treated as pretty paths for proper processing (@dharezlak)
 - JS resources required for the OFF viewer are loaded in the file browser direct embed mode (@dharezlak)
 - Resource pretty path decodes and encodes values according to URI restrictions (@dharezlak)
+- Local path exclusion validation query fixed (@ddharezlak)
 
 ### Security