+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 - Move FileStore configuration into one place (@mkasztelnik)
 - Make patient `case_number` factory field unique (@mkasztelnik)
 - Redirect user back to previous page after proxy is regenerated successfully (@mkasztelnik)
+- Prefix is removed from segmentation outputs (@mkasztelnik)
 
 ### Deprecated