+++ CHANGELOG.md
@@ -41,6 +41,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 - Set Rimrock computation job id to nil while restarting computation (@mkasztelnik)
 - Show missing patient outputs on pipeline diff view (@mkasztelnik)
 - Path processing fixed for policy move/copy operations invoked via API (@dharezlak)
+- Set correct order for pipelines computations (@mkasztelnik)
 
 ### Security