+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 Please view this file on the master branch, on stable branches it's out of date.
 
-## [Unreleased]
+## 0.5.0
 
 ### Added
 - Pipelines have own DataFiles which represent produced results (@Nuanda)